src/requirements.txt
pillow>=9.0.0
pillow-heif>=0.10.0
pyheif>=0.7.0
piexif
